Slight GUI improvement for MacOS.

+++ src/albumish/PlayerThread.java
@@ -306,7 +306,7 @@ public class PlayerThread extends Thread {
      * Tell the player to jump forwards or backwards in the audio stream.
      */
     public void do_audio_jump(float value) {
-        this.target_position = Float.valueOf(value);
+        this.target_position = value;
     }
 
     /**
